Fixes expected path in tests for get_executable().
On Fedora, the `ls' executable is at /usr/bin/ls instead of /bin/ls. The
test code now runs `which ls' to determine what the expected value should be.
(Which is probably what I should have written initially anyway, since the
point of get_executable() was to mimic the behavior of `which'.)