Formally drops support for Python 3.3.
This, like dropping support for Python 2.6, was a long time
coming. Recent Travis CI builds have been failing, so that was
the final straw. Safe is designed to be run on the daily drivers
of technically competent folks, so the machine should already
have a sufficient Python version and if not it's safe to assume
the user can install one.